use colors!
